+++ src/haas_mqtt_new.c
@@ -573,7 +573,7 @@ for(int i =0;i<haas_device_num;i++)
 			                "\t\"V0%d\": \"%s\",\r\n", i + 1, dev->value_text);
 		} else {
 			len1 = snprintf(s_data + len, remaining,
-			                "\t\"V0%d\": %.1f,\r\n", i + 1, dev->value2);
+			                "\t\"V0%d\": %d,\r\n", i + 1, (int)dev->value2);
 		}
 	}
 	else
@@ -586,7 +586,7 @@ for(int i =0;i<haas_device_num;i++)
 			                "\t\"V%d\": \"%s\",\r\n", i + 1, dev->value_text);
 		} else {
 			len1 = snprintf(s_data + len, remaining,
-			                "\t\"V%d\": %.1f,\r\n", i + 1, dev->value2);
+			                "\t\"V%d\": %d,\r\n", i + 1, (int)dev->value2);
 		}
 	}
 	if (len1 < 0) {
