+++ src/mqtt.c
@@ -542,7 +542,7 @@ for(int i =0;i<haas_device_num;i++)
 			                "\t\"V0%d\": \"%s\",\r\n", i + 1, dev->value_text);
 		} else {
 			len1 = snprintf(s_data + len, remaining,
-			                "\t\"V0%d\": %.3f,\r\n", i + 1, dev->value2);
+			                "\t\"V0%d\": %d,\r\n", i + 1, (int)dev->value2);
 		}
 	}
 	else
@@ -555,7 +555,7 @@ for(int i =0;i<haas_device_num;i++)
 			                "\t\"V%d\": \"%s\",\r\n", i + 1, dev->value_text);
 		} else {
 			len1 = snprintf(s_data + len, remaining,
-			                "\t\"V%d\": %.3f,\r\n", i + 1, dev->value2);
+			                "\t\"V%d\": %d,\r\n", i + 1, (int)dev->value2);
 		}
 	}
 	if (len1 < 0) {
@@ -578,10 +578,10 @@ if (g_energy_window_value_ready) {
 	if (remaining > 1) {
 	if (dev_num < 10) {
 		len1 = snprintf(s_data + len, remaining,
-		                "\t\"V0%d\": %.1f,\r\n", dev_num, g_energy_window_value_wh);
+		                "\t\"V0%d\": %d,\r\n", dev_num, (int)g_energy_window_value_wh);
 	} else {
 		len1 = snprintf(s_data + len, remaining,
-		                "\t\"V%d\": %.1f,\r\n", dev_num, g_energy_window_value_wh);
+		                "\t\"V%d\": %d,\r\n", dev_num, (int)g_energy_window_value_wh);
 	}
 	if (len1 < 0) {
 		s_data[len] = '\0';
